backend/requirements.txt
# Student RAG Assistant — Python dependencies
# Install: pip install -r requirements.txt

fastapi==0.115.5
uvicorn[standard]==0.32.1
python-multipart==0.0.12
python-dotenv==1.0.1

# LangChain
langchain==0.3.7
langchain-community==0.3.7
langchain-groq==0.2.1

# PDF
pypdf==5.1.0

# Local embeddings + reranker (CPU, no API key needed)
# all-MiniLM-L6-v2              ~90MB  — embedding model
# cross-encoder/ms-marco-MiniLM-L-6-v2  ~80MB  — reranker model
# Both download automatically on first use, then cached
sentence-transformers==3.3.1

# Vector store (local, no server)
chromadb==0.5.20

# Groq SDK (for catching RateLimitError)
groq>=0.9.0

pydantic==2.10.2

# Database — chat history persistence
sqlalchemy==2.0.35
alembic==1.14.0

# Auth — JWT tokens + password hashing
python-jose[cryptography]==3.3.0
bcrypt==4.2.0
email-validator==2.2.0

# Updated LangChain integrations (fixes deprecation warnings)
langchain-huggingface==0.1.2
langchain-chroma==0.1.4

# Rate limiting
slowapi==0.1.9

# Testing
pytest==8.3.4
pytest-asyncio==0.24.0
httpx==0.27.0
psycopg2-binary